requirements.txt
langchain==0.0.162
openai==0.27.7
tiktoken==0.4.0
pinecone-client[grpc]==2.2.1 # pip install "pinecone-client[grpc]"
pinecone_datasets #=='0.5.0rc10'